fix: preserve valid legacy settings during normalization
