+++ server/storage.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from "crypto";
 import fs from "fs/promises";
 import path from "path";
-import { generalSettingsSchema } from "@shared/schema";
+import { createDefaultGeneralSettings, getGeneralSettingCategory, normalizeGeneralSettings } from "@shared/generalSettings";
 import type {
   UrlRule,
   InsertUrlRule,
@@ -50,179 +50,6 @@ const URL_TRACKING_SORT_COLUMNS = new Set([
   "matchQuality",
 ]);
 
-
-const GENERAL_SETTING_CATEGORIES: Record<string, string> = {
-  id: "metadata",
-  updatedAt: "metadata",
-  headerTitle: "header",
-  headerIcon: "header",
-  headerLogoUrl: "header",
-  headerBackgroundColor: "header",
-  popupMode: "display",
-  mainTitle: "main_content",
-  mainDescription: "main_content",
-  mainBackgroundColor: "main_content",
-  alertIcon: "alert",
-  alertBackgroundColor: "alert",
-  urlComparisonTitle: "url_comparison",
-  urlComparisonIcon: "url_comparison",
-  urlComparisonBackgroundColor: "url_comparison",
-  oldUrlLabel: "url_comparison",
-  newUrlLabel: "url_comparison",
-  defaultNewDomain: "redirect_defaults",
-  defaultRedirectMode: "redirect_defaults",
-  defaultSearchUrl: "redirect_defaults",
-  defaultSearchMessage: "redirect_defaults",
-  defaultSearchSkipEncoding: "redirect_defaults",
-  smartSearchRegex: "smart_search",
-  smartSearchRules: "smart_search",
-  enableCopyButton: "actions",
-  enableOpenButton: "actions",
-  newUrlClickBehavior: "actions",
-  copyButtonText: "actions",
-  openButtonText: "actions",
-  showUrlButtonText: "actions",
-  popupButtonText: "actions",
-  specialHintsTitle: "special_hints",
-  specialHintsDescription: "special_hints",
-  specialHintsIcon: "special_hints",
-  infoTitle: "additional_info",
-  infoTitleIcon: "additional_info",
-  infoItems: "additional_info",
-  infoIcons: "additional_info",
-  footerCopyright: "footer",
-  caseSensitiveLinkDetection: "matching",
-  encodeImportedUrls: "imports",
-  autoRedirect: "redirect_defaults",
-  showLinkQualityGauge: "quality",
-  matchHighExplanation: "quality",
-  matchMediumExplanation: "quality",
-  matchLowExplanation: "quality",
-  matchRootExplanation: "quality",
-  matchNoneExplanation: "quality",
-  enableTrackingCache: "tracking",
-  maxStatsEntries: "tracking",
-  enableReferrerTracking: "tracking",
-  enableFeedbackSurvey: "feedback",
-  feedbackSurveyTitle: "feedback",
-  feedbackSurveyQuestion: "feedback",
-  feedbackSuccessMessage: "feedback",
-  feedbackButtonYes: "feedback",
-  feedbackButtonNo: "feedback",
-  enableFeedbackComment: "feedback",
-  feedbackCommentTitle: "feedback",
-  feedbackCommentDescription: "feedback",
-  feedbackCommentPlaceholder: "feedback",
-  feedbackCommentButton: "feedback",
-  enableFeedbackSmartSearchFallback: "feedback",
-  feedbackSmartSearchFallbackTitle: "feedback",
-  feedbackSmartSearchFallbackDescription: "feedback",
-  feedbackSmartSearchFallbackQuestion: "feedback",
-  showSatisfactionTrend: "analytics",
-  satisfactionTrendFeedbackOnly: "analytics",
-  satisfactionTrendDays: "analytics",
-  globalSearchAndReplace: "global_rules",
-  globalStaticQueryParams: "global_rules",
-  globalKeptQueryParams: "global_rules",
-};
-
-function createDefaultGeneralSettings(): GeneralSettings {
-  return {
-    id: randomUUID(),
-    headerTitle: "URL Migration Tool",
-    headerIcon: "ArrowRightLeft",
-    headerBackgroundColor: "#ffffff",
-    popupMode: "active",
-    mainTitle: "Veralteter Link erkannt",
-    mainDescription: "Sie verwenden einen veralteten Link unserer Web-App. Bitte aktualisieren Sie Ihre Lesezeichen und verwenden Sie die neue URL unten.",
-    mainBackgroundColor: "#ffffff",
-    alertIcon: "AlertTriangle",
-    alertBackgroundColor: "yellow",
-    urlComparisonTitle: "URL-Vergleich",
-    urlComparisonIcon: "ArrowRightLeft",
-    urlComparisonBackgroundColor: "#ffffff",
-    oldUrlLabel: "Alte URL (veraltet)",
-    newUrlLabel: "Neue URL (verwenden Sie diese)",
-    defaultNewDomain: "https://thisisthenewurl.com/",
-    defaultRedirectMode: "domain",
-    defaultSearchUrl: null,
-    defaultSearchMessage: "Keine direkte Übereinstimmung gefunden. Sie werden zur Suche weitergeleitet.",
-    defaultSearchSkipEncoding: false,
-    smartSearchRegex: null,
-    smartSearchRules: [],
-    enableCopyButton: true,
-    enableOpenButton: true,
-    newUrlClickBehavior: "copy",
-    copyButtonText: "URL kopieren",
-    openButtonText: "In neuem Tab öffnen",
-    showUrlButtonText: "Zeige mir die neue URL",
-    popupButtonText: "Zeige mir die neue URL",
-    specialHintsTitle: "Spezielle Hinweise für diese URL",
-    specialHintsDescription: "Hier finden Sie spezifische Informationen und Hinweise für die Migration dieser URL.",
-    specialHintsIcon: "Info",
-    infoTitle: "Zusätzliche Informationen",
-    infoTitleIcon: "Info",
-    infoItems: ["", "", ""],
-    infoIcons: ["Bookmark", "Share2", "Clock"],
-    footerCopyright: "© 2024 URL Migration Service. Alle Rechte vorbehalten.",
-    caseSensitiveLinkDetection: false,
-    encodeImportedUrls: true,
-    autoRedirect: false,
-    showLinkQualityGauge: true,
-    matchHighExplanation: "Die neue URL entspricht exakt der angeforderten Seite oder ist die Startseite. Höchste Qualität.",
-    matchMediumExplanation: "Die URL wurde erkannt, weicht aber leicht ab (z.B. zusätzliche Parameter).",
-    matchLowExplanation: "Es wurde nur ein Teil der URL erkannt und ersetzt (Partial Match).",
-    matchRootExplanation: "Startseite erkannt. Direkte Weiterleitung auf die neue Domain.",
-    matchNoneExplanation: "Die URL konnte nicht spezifisch zugeordnet werden. Es wird auf die Standard-Seite weitergeleitet.",
-    enableTrackingCache: true,
-    maxStatsEntries: 0,
-    enableReferrerTracking: true,
-    enableFeedbackSurvey: false,
-    feedbackSurveyTitle: "War die neue URL korrekt?",
-    feedbackSurveyQuestion: "Dein Feedback hilft uns, die Weiterleitungen weiter zu verbessern.",
-    feedbackSuccessMessage: "Vielen Dank für deine Rückmeldung.",
-    feedbackButtonYes: "Ja, OK",
-    feedbackButtonNo: "Nein",
-    enableFeedbackComment: false,
-    feedbackCommentTitle: "Kennen Sie die korrekte URL?",
-    feedbackCommentDescription: "Bitte geben Sie die korrekte URL hier ein, damit wir sie korrigieren können.",
-    feedbackCommentPlaceholder: "https://...",
-    feedbackCommentButton: "Absenden",
-    enableFeedbackSmartSearchFallback: false,
-    feedbackSmartSearchFallbackTitle: "Vorschlag: Suche verwenden",
-    feedbackSmartSearchFallbackDescription: "Keine passende Weiterleitung gefunden. Versuchen Sie es mit der Suche.",
-    feedbackSmartSearchFallbackQuestion: "Hat dieser Link funktioniert?",
-    showSatisfactionTrend: true,
-    satisfactionTrendFeedbackOnly: false,
-    satisfactionTrendDays: 30,
-    globalSearchAndReplace: [],
-    globalStaticQueryParams: [],
-    globalKeptQueryParams: [],
-    updatedAt: new Date().toISOString(),
-  };
-}
-
-function getGeneralSettingCategory(key: string): string {
-  return GENERAL_SETTING_CATEGORIES[key] ?? "custom";
-}
-
-function normalizeGeneralSettings(settings: Partial<GeneralSettings>, id: string): GeneralSettings {
-  const defaults = createDefaultGeneralSettings();
-  const knownSettings = Object.fromEntries(
-    Object.entries(settings).filter(([key]) => key in GENERAL_SETTING_CATEGORIES),
-  ) as Partial<GeneralSettings>;
-  const normalized = {
-    ...defaults,
-    ...knownSettings,
-    id,
-    smartSearchRules: knownSettings.smartSearchRules ?? (
-      knownSettings.smartSearchRegex ? [{ pattern: knownSettings.smartSearchRegex, order: 0 }] : defaults.smartSearchRules
-    ),
-    updatedAt: knownSettings.updatedAt ?? defaults.updatedAt,
-  };
-
-  return generalSettingsSchema.parse(normalized);
-}
 
 function normalizeSortOrder(sortOrder: string | undefined): "ASC" | "DESC" {
   return sortOrder?.toLowerCase() === "asc" ? "ASC" : "DESC";
